cheri: wireless: intel: Fix uintptr_t use
- Align structures that contain pointers.
- Add uintptr_t casts.

There is one cast that looks like an upstream bug:
"data->replay_ctr" is of type "u8 *" which suggests that it
already _points_ to the replay counter. However, when passing
the argument to be64_to_cpup the address of data->replay_ctr
is taken which apparently results in the _pointer_ being endian
converted and not the data it points to.

+++ drivers/net/wireless/intel/iwlegacy/common.h
@@ -574,7 +574,7 @@ struct il_device_cmd_huge {
 
 struct il_host_cmd {
 	const void *data;
-	unsigned long reply_page;
+	uintptr_t reply_page;
 	void (*callback) (struct il_priv *il, struct il_device_cmd *cmd,
 			  struct il_rx_pkt *pkt);
 	u32 flags;
@@ -1497,7 +1497,7 @@ __il_free_pages(struct il_priv *il, struct page *page)
 }
 
 static inline void
-il_free_pages(struct il_priv *il, unsigned long page)
+il_free_pages(struct il_priv *il, uintptr_t page)
 {
 	free_pages(page, il->hw_params.rx_page_order);
 	il->alloc_rxb_page--;
